requirements.txt
openai>=2.8.1
python-dotenv>=1.0.0
